- Rename files - Use Camel case for variables - cleanup namespaces
